Set matplotlib backend during initialization
On Debian 12, an error was caused  when trying to plot the graphs in
both the voaAreaPlot.py and voaP2PPlot.py ("ImportError: Cannot load
backend 'TkAgg' which requires the 'tk' interactive framework, as 'gtk3'
is currently running").

This patch initializes the backend for matplotlib explicitly during
object initialization, avoiding the problem.