Add missing manpages
There were missing manpages to some of the software executables. Some
distributions, such as Debian, consider it a bug not having manpages to
programs that are not auxiliary ones.

This commit provides the same manpages that were written for the Debian
packaging of pythonprop, in the hope they will be useful as a starting
point for an official manpage (or used "as is", if deemed so), if there
is interest in implementing them.